+++ prettier.config.js
@@ -1,8 +1,9 @@
 import { createConfig } from '@alexlit/config-prettier';
 
 export default createConfig({
+  jsdoc: true,
   php: true,
   ruby: true,
   sql: true,
-  tailwindcss: true,
+  tailwindcss: false,
 });