+++ stylelint.config.js
@@ -1,8 +1,9 @@
 import { createConfig } from '@alexlit/config-stylelint';
 
 export default createConfig({
-  a11y: true,
-  gamut: true,
+  'a11y': true,
+  'bem': true,
+  'gamut': true,
   'logical-css': true,
   'media-use-custom-media': true,
   'no-indistinguishable-colors': false,